chore(daemon): relocate perm config
